updated Postman json for OMS, CMS, SMS, and PMS

+++ ECom-app/App/src/main/java/com/itsthatjun/ecommerce/mbg/model/OrderItem.java
@@ -24,7 +24,7 @@ public class OrderItem implements Serializable {
 
     private Integer productQuantity;
 
-    private Long productSkuId;
+    private Integer productSkuId;
 
     private String productSkuCode;
 
@@ -120,11 +120,11 @@ public class OrderItem implements Serializable {
         this.productQuantity = productQuantity;
     }
 
-    public Long getProductSkuId() {
+    public Integer getProductSkuId() {
         return productSkuId;
     }
 
-    public void setProductSkuId(Long productSkuId) {
+    public void setProductSkuId(Integer productSkuId) {
         this.productSkuId = productSkuId;
     }
